refactored data/, made dataset float32-only

+++ deeplearning/classifiers/mlp.py
@@ -250,7 +250,7 @@ class OpMLPPredict(OpPredict, Classification, Regression):
         b = roi.stop[0]
 
         inputs = self.Input[a:b, ...].wait()
-        # inputs = inputs.astype(np.float32)
+        inputs = inputs.astype(np.float32)
         shared = theano.shared(inputs, name='inputs')
         prediction = model.fprop(shared).eval()
